refactor: simplify navigator and auth flow
