+++ src/presentation/navigation/navigators/app-navigator.tsx
@@ -1,76 +1,22 @@
 import React, { useContext } from 'react';
 import { NavigationContainer, DefaultTheme } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { routes } from '~/presentation/navigation/routes';
+
+// Navigators
+import { AuthNavigator } from './auth-navigator';
+import { PrivateNavigator } from './private-navigator';
 
 // Components
-import { AppointmentsIcon, HomeIcon } from '~/presentation/components/icons';
 import { ActivityIndicator, View } from 'react-native';
-
-// Screens
-import {
-  AppointmentDetailsScreen,
-  AppointmentsScreen,
-  CreateAppointmentScreen,
-  LoginScreen,
-} from '~/presentation/screens';
 
 // Styles
 import { useTheme } from 'styled-components';
 
 // Types
-import { RootStackParamList } from './types';
 import { AuthContext } from '~/presentation/context';
 
 export function AppNavigator() {
   const { isLoading, user } = useContext(AuthContext);
-  const Stack = createNativeStackNavigator<RootStackParamList>();
-  const Tab = createBottomTabNavigator();
   const theme = useTheme();
-
-  function renderAuthFlow() {
-    if (user?.token) {
-      return null;
-    }
-    return (
-      <Stack.Screen
-        name={routes.LoginScreen}
-        component={LoginScreen}
-        options={{ headerShown: false }}
-      />
-    );
-  }
-
-  function HomeScreen() {
-    return (
-      <Tab.Navigator
-        initialRouteName={routes.CreateAppointmentScreen}
-        screenOptions={{
-          headerShown: false,
-          tabBarActiveTintColor: theme.colors.secondary,
-          tabBarInactiveTintColor: theme.colors.black,
-        }}
-      >
-        <Tab.Screen
-          name={routes.CreateAppointmentScreen}
-          component={CreateAppointmentScreen}
-          options={{
-            title: 'Home',
-            tabBarIcon: ({ color }) => <HomeIcon color={color} />,
-          }}
-        />
-        <Tab.Screen
-          name={routes.AppointmentsScreen}
-          component={AppointmentsScreen}
-          options={{
-            title: 'Consultas',
-            tabBarIcon: ({ color }) => <AppointmentsIcon color={color} />,
-          }}
-        />
-      </Tab.Navigator>
-    );
-  }
 
   const screensTheme = {
     ...DefaultTheme,
@@ -90,21 +36,7 @@ export function AppNavigator() {
 
   return (
     <NavigationContainer theme={screensTheme}>
-      <Stack.Navigator>
-        {renderAuthFlow()}
-
-        <Stack.Screen
-          name={routes.HomeScreen}
-          component={HomeScreen}
-          options={{ headerShown: false }}
-        />
-
-        <Stack.Screen
-          name={routes.AppointmentDetailsScreen}
-          component={AppointmentDetailsScreen}
-          options={{ title: 'Detalhes' }}
-        />
-      </Stack.Navigator>
+      {!user ? <AuthNavigator /> : <PrivateNavigator />}
     </NavigationContainer>
   );
 }
